Maik I emailed you but this change is bugging me so much. Hey this is why we have CVS right?
From what I can see in no way should the page returned be based upon the users request. Further if a user does a POST this will return the page "POST", which is meaningless. If for some god knows why reason request isnt present this will return context which is an unpack error.

Those things combined is why Im removing this patch, we should write a unit test and make sure these changes were made for the right reason.