Plip 80 - Nav Tree refactoring
With use of ExtendedPathIndex the creation of the nav tree will be much less
consuming. By refactoring the code of NavigationPortlet (part of PlonePortlets)
All NavTreeBuilder stuff becomes obsolete.

This is a first checking with just the refactoring. Integration of the existing properties will be done by jladage and joelburton.

- Added createNavTree method in PloneTool.py

- Deleted setupForms method from Portal.py

- Added setupNavTree method in Portal.py which:

   - sets up the navtree propertysheet
   
   - replaces the path index with an ExtendedPathIndex typed index

   - creates the necessary indexes
 
- Added ExtendedPathIndex initialization in PloneTestCase.py

- Updated History.

- Added some scripts