Merged (manually - the original branch was so outdated it was much easier to do it by hand) PLIP76 - author feedback. This makes the name in the document_byline a link (so long as the author is a member managed by the portal) to a page (author.cpt) from which author info can be obtained, and feedback can be sent. The template needs UI work, definitely.
A new method in MembershipTool.py was added (getMemberInfo) for getting "safe"
member info to display on this page (getMemberById is protected by manager
permissions). Currently it returns only the full name, but we should add more
member data (e.g. location, phone number, etc.) and then add them to the
dict returned by this method, so that they can be placed on the author page.