Commiting (hopefully) working implementation of PLIP97 - getting the "display" menu at the portal root.
The implementation provides a mixin class PropertyManagedBrowserDefault, which
picks the list of selectable views from a property called selectable_views.
This is mixed into the portal object. By contrast, the ATCT implementation
of the same interface (ISelectableBrowserDefault) replies on TemplateMixin
from AT, which in turns holds a registry of types in archetype_tool. The
interface (and hence Plone) doesn't care.

A couple of weaknesses were discovered and rectified in the way the interface
operated though. Notably, canSetLayout() used to do a check whether there
was more than one template to select and return false if there was only one.
This has now been moved to the higher level getSelectableViews.py, where it
belongs, since it's mostly a user interface issue. Also, getAvailableLayouts()
used to return a DisplayList, which is AT-only; it now returns a list of
tuples (id, title).

NOTE: This implies a small change in ATContentTypes, as it used to returns
a DisplayList. I will commit that as soon as this branch is merged.

Note that the implementation