Checking in the extra skin support in #4546
If you select skins using the changeSkin method RR does not seem to notice the change. There are also various problems with cacheing of the wrong 
skin if you change skin or use different skins for different sections of your site.

This patch changes the links that RR creates from:
<portal_url>/portal_{css|javascripts}/resource
to:
<portal_url>/portal_{css|javascripts}/<skin_name>/resource

When the resource is fetched the <skin_name> traversal step changes the skin for that request to skin_name before returning the resource. if you 
link to <portal_url>/portal_{css|javascripts}/resource it should work as it does currently (as long as your resource is not the same name as a 
skin name).

It would be nice to make the skin awareness optional on the tool (for javascripts this may be better). This should be relatively straight forward, 
just need to make sure that the header macros check for this.