Making the full name of the user show up in the personal bar; this fixes #5268.
Somebody, can we please, please, please add a convenience method on the relevant
tool to get the full name with fallback to username if there is none?

Doing the following dance every time you want to get the full name of the person
in a template is insane:

<p tal:define="author python:mtool.getMemberInfo(user.getId())">
  <span tal:replace="python:author and author['fullname'] or user.getId()">
  Fullname with fallback to username
  </span>
</p>

Poi has a method getNiceName() that does this, I think. Backport? :)