Fix Five resource handling. Request and response headers must be handled specially.
The GET for the z3 resource tries to handle a request autonomously, especially it
can happen that if the browser has put If-Modified-Since on the request, it
would give a 30x response status. This, if stays on the response, causes to
hang the browser indefinitely while loading the page. Solution is clear the
header and restore the response headers after the call to GET.