Move some of the generic navigation elements out of CMFPlone.browser and into here, specifically:
 - INavigationRoot
 - IDefaultPage
 - INavtreeQueryBuilder
 - INavtreeStrategy

And also the generic navtree helper functions. The idea is that 
Products.CMFPlone.browser shouldn't be a dumping ground for all kind of
UI related interfaces and infrastructure, so we move it here, and let
the CMFPlone browser package be concerned with actual views and policy
decisions that directly affect Plone's look-and-feel.

Note - tests are still in CMFPlone :)