A much cleaner table-sort with jquery.
Note that we no longer use a comparison function but rather build up the sorting array with cleaned-up data instead. This should be much faster; the comparison method didn't do anything but convert types anyway.

Second note is that it introduces a new classname on the sorting arrow images to make it easy to select them.

Last is the use of an attribute on the table to store column sorting state (to detect reverse). The old version checked if the first 2 lines of the table were already in sorting order to decide to reverse.