Updating the flag/language listings.
The following logic is used:

- If a language is only used in one country, then that flag is mapped to it.
  Example: Finnish (fi and fi-fi)

- If a language is the *only* official language in a country, then that
  country's flag is used. Example: Chile (es-cl) - language is Spanish, but 
  it's the only official language). This is the most common configuration
  when you are using combined language codes (xx-yy), but there are some 
  exceptions.

- If there is more than one official language, then the flags reflect the 
  origin of the language, which may or may not be acceptable, always check
  this. Example: Swedish Finnish (sv-fi) is one of the official languages in
  Finland, but will show the Swedish flag, so the difference can be discerned
  when both fi(-fi) and sv-fi is used with flags. Usually we recommend using
  textual links in these cases, but if you are 100% sure that flags are OK,
  this is what will happen.

There might still be mappings that are wrong, but we have tried to check as 
many of them as possible. If you find any problems, please contact the 
Internationalization list at http://plone.org/support.

Remember, this makes an assumption that flags are a politically 
non-controversial way of representing language choice in that area - you still
need to check if use of flags to represent languages is inappropriate in your context. If in doubt, always use textual links instead of flags.