the hotfix will need to have a genericsetup profile in order to override the critical templates from CMFPlone.  however, the profile cannot be loaded from it's `initialize`, except perhaps if it were to search for all plone sites within the zope instance.  so, installing the hotfix will not only require to extract it into `Products/`, but also to quick-install it afterwards.
to avoid having the user quick-install two profiles, i.e. one for `plone.app.protect` _and_ one for the hotfix itself, the existing profile and `configure.zcml` from the former gets copied.  this not only makes the setup easier, but also allows to add the modified templates, which are (and should not be) part of `plone.app.protect`.