New branch to add support for multiple catalogs.
This is needed to support membrane, which requires that membrane-using
objects also be stored in the ``membrane_tool`` catalog.

In addition we'll try to get a useable UID for dexterity content. Currently
the UID index in ``portal_catalog`` is acquired for dexterity content, which
will lead to duplicate UIDs in the catalog.