* Properly handle nesting of groups. UI now allows addition and display of groups within other groups. Closes #8556 * @@usergroup-groupprefs now shows an icon to designate that the group has inherited that global role from another group.
For posterity's sake: We've made the design decision that groups that have an explicitly-assigned global role *and* inherit same role from a parent group will display the "inherited" icon (right now, the Plone logo used by @@sharing). The question of "how would I assign/unassign a explicit global role on something that inherits the same?" did come up, but we felt that this was an edge case and wasn't vital enough to confuse the UI for the large majority of users. Removing the group from its parent group (the one assigning the inherited role), explicitly assigning that role, and then readding should provide a somewhat convoluted, but approachable workaround.